Attempt #2 at fixing the TP handler.
OnInteract doesn’t act immediately because it allows for a long-press. If the digit changes while the button is being held for .1 seconds, the user gets a strike even if they gave the right answer. Therefore, use eXish’s solution to wait for it to come _off_ the digit first before waiting to get back on it so we know we have enough time to press the button.